+++ core/buffers/kernels/spatialGradient.cpp
@@ -61,18 +61,20 @@ G12Buffer* SpatialGradient::findCornerPoints(double scaler, int apperture)
 
         blurerV.process(&start   , &blurVert);
         blurerH.process(&blurVert, &blur);
-    } else {
+    }
+    else
+    {
         printf("SpatialGradient::findCornerPoints() %d\n", v);
 
-        for(int i = 0 ; i < start->h; i++)
+        for (uint32_t i = 0 ; i < start->h; i++)
         {
-            for(int  j = 0; j < start->w; j++)
+            for (uint32_t j = 0; j < start->w; j++)
             {
                 Vector3dd sum = Vector3dd::Zero();
 
-                for(int dy = i - v / 2 ; dy <= i + v / 2; dy++)
+                for (int dy = i - v / 2 ; dy <= (int)i + v / 2; dy++)
                 {
-                    for(int dx = j - v / 2 ; dx <= j + v / 2; dx++)
+                    for (int dx = j - v / 2 ; dx <= (int)j + v / 2; dx++)
                     {
                         if (!start->isValidCoord(dy,dx))
                             continue;
